requirements.txt
opencv-python
pillow
requests
numpy
huggingface-cli